Remove `td` and `table` css classes
Summary:
CSS style rules applied to raw DOM elements (`td` and `table`) are a bad idea and in this case are breaking some core Wordpress functionality (the post scheduling calendar). As it seems like there are no longer any `td` nor `table` elements involved in the Chat plugin, nor in the WP wrapper, this diff removes the style rules entirely.

**Before:**

{F348085242}

**After:**

{F348085317}

Reviewed By: crazycoder26

Differential Revision: D25290815

fbshipit-source-id: b51586853ba9c889889eb8b0c79328367b27db77