Add review prompt notice to admin dashboard
Summary:
Show a notice card in the admin dashboard, prompting admin users to leave a review for the Facebook Chat plugin.

Includes dependency third-party extension (`persist-admin-notices-dismissal`) for creating a dismissal record so that the notice does not persist on subsequent sessions after being dismissed.

This diff does not include any gating for the notice.

{F628312421}

Reviewed By: alanhhwong

Differential Revision: D29418414

fbshipit-source-id: d08c91bd3207021257ddfa351f3cd02ba235cdc8