Fix Advanced Configuration panel so that all pages are selected by default for "custom" setting
Summary:
Previously, when accessing the "Custom pages" option in "Advanced Configuration" for the first time, **all** the pages of your site would be unchecked by default, meaning that you'd have to choose every page to include, rather than simply selecting pages to exclude.

This could also cause big problems if an admin experimented with switching to "Custom pages", but did not update any of the options (this would result in a setting in which the Chat plugin would not be shown on **any** pages!).

This diff rewrites the checks when rendering the options page, to check all boxes by default, if no previous settings exist.

{F628770403}

Reviewed By: alanhhwong

Differential Revision: D29455117

fbshipit-source-id: 09b98fcbda21940920b73844aedb134ac3d5b81c