Render commented line of markup when Chat plugin disabled in advanced config
Summary:
With the addition of the "Advanced Configuration" panel in v2.1 of the Wordpress plugin, Wordpress admins can now control exactly which pages of their site the Chat plugin will load on. However, this can make debugging trickier, as if the Chat plugin code is not present on a page, there is no way to determine, from the outside, whether this is due to an issue with the Wordpress plugin itself, or whether the admin has simply set the plugin not to show on that particular page.

This diff adds a line of commented code when the WP plugin's configuration is preventing the Chat plugin from loading. The commented string is "fbmcc-config-disabled" (but could theoretically be anything reasonably unique and distinguishable).

Reviewed By: alanhhwong

Differential Revision: D29454272

fbshipit-source-id: 76aeaa541dc630d7b90af04e1920015331fd94fe